LBFGSLite

Documentation for LBFGSLite.

LBFGSLite.LBFGSParams

LBFGSLite.optimize

LBFGSLite.LBFGSParams — Type

struct LBFGSParams{T <: AbstractFloat}

Parameter struct for LBFGS.

Arguments

mem_size::Int = 8: The number of corrections to approximate the inverse hessian matrix. The L-BFGS routine stores the computation results of previous m iterations to approximate the inverse hessian matrix of the current iteration. This parameter controls the size of the limited memories (corrections). The default value is 8. Values less than 3 are not recommended. Large values will result in excessive computing time.

g_epsilon::T = 1.0e-5: Epsilon for grad convergence test. DO NOT USE IT in nonsmooth cases! Set it to 0.0 and use past-delta-based test for nonsmooth functions. This parameter determines the accuracy with which the solution is to be found. A minimization terminates when ||g(x)||inf / max(1, ||x||inf) < gepsilon, where ||.||inf is the infinity norm. The default value is 1.0e-5. This should be greater than 1.0e-6 in practice because L-BFGS does not directly reduce first-order residual. It still needs the function value which can be corrupted by machine_prec when ||g|| is small.

past::Int = 3: Distance for delta-based convergence test. This parameter determines the distance, in iterations, to compute the rate of decrease of the cost function. If the value of this parameter is zero, the library does not perform the delta-based convergence test. The default value is 3.

delta::T = 1.0e-6: Delta for convergence test. This parameter determines the minimum rate of decrease of the cost function. The library stops iterations when the following condition is met: |f' - f| / max(1, |f|) < delta, where f' is the cost value of past iterations ago, and f is the cost value of the current iteration. The default value is 1.0e-6.

max_iterations::Int = 0: The maximum number of iterations. The lbfgsoptimize() function terminates an minimization process with ::LBFGSERRMAXIMUMITERATION status code when the iteration count exceedes this parameter. Setting this parameter to zero continues an minimization process until a convergence or error. The default value is 0.

max_linesearch::Int = 64: The maximum number of trials for the line search. This parameter controls the number of function and gradients evaluations per iteration for the line search routine. The default value is 64.

min_step::T = 1.0e-20: The minimum step of the line search routine. The default value is 1.0e-20. This value need not be modified unless the exponents are too large for the machine being used, or unless the problem is extremely badly scaled (in which case the exponents should be increased).

max_step::T = 1.0e+20: The maximum step of the line search. The default value is 1.0e+20. This value need not be modified unless the exponents are too large for the machine being used, or unless the problem is extremely badly scaled (in which case the exponents should be increased).

f_dec_coeff = 1.0e-4: A parameter to control the accuracy of the line search routine. The default value is 1.0e-4. This parameter should be greater than zero and smaller than 1.0.

s_curv_coeff = 0.9: A parameter to control the accuracy of the line search routine. The default value is 0.9. If the function and gradient evaluations are inexpensive with respect to the cost of the iteration (which is sometimes the case when solving very large problems) it may be advantageous to set this parameter to a small value. A typical small value is 0.1. This parameter should be greater than the fdeccoeff parameter and smaller than 1.0.

cautious_factor = 1.0e-6: A parameter to ensure the global convergence for nonconvex functions. The default value is 1.0e-6. The parameter performs the so called cautious update for L-BFGS, especially when the convergence is not sufficient. The parameter must be positive but might as well be less than 1.0e-3 in practice.

LBFGSLite.optimize — Function

optimize(
    fg!::F
    x::AbstractVector{T},
    params::LBFGSParams{T} = LBFGSParams{T}(),
) = optimize!(fg!, copy(x), LBFGSWorkspace(x, params), params)

optimize!(
    fg!::F
    x::AbstractVector{T},
    work::LBFGSWorkspace{T},
    params::LBFGSParams{T} = LBFGSParams{T}(),
) -> x, fx::T, ls::LBFGS_STATUS

Minimize a function using L-BFGS.

Assumptions:

f(x) is either C2 or C0 but piecewise C2;

f(x) is lower bounded;

f(x) has bounded level sets;

g(x) is either the gradient or subgradient;

The gradient exists at the initial guess x0.

Arguments

fg!(x, g)::F: In-place function that computes the objective fx = fg!(x, g) and stores the gradient in g.

x::AbstractVector{T}: Vector of decision variables. Used for the initial guess.

work::LBFGSWorkspace{T}: Struct with pre-allocated arrays.

params::LBFGSParams{T} = LBFGSParams{T}(): The parameters for L-BFGS optimization.

Returns

x::AbstractVector{T}: Vector of decision variables.

fx::T: Final value of cost function.

ls::LBFGS_STATUS: Status code.